refactor: move routes to the api package
