+++ controlplane/backend/internal/api/desiredstate/desiredstate.go
@@ -22,7 +22,7 @@ package desiredstate
 import (
 	"github.com/gin-gonic/gin"
 
-	"github.com/hoophq/hoop/controlplane/backend/internal/apierr"
+	"github.com/hoophq/hoop/controlplane/backend/internal/api/apierr"
 )
 
 const ticket = "EVL-231"
