+++ controlplane/backend/internal/api/health.go
@@ -1,4 +1,4 @@
-package httpapi
+package api
 
 import (
 	"context"
