+++ controlplane/backend/internal/api/inventory/inventory.go
@@ -23,7 +23,7 @@ package inventory
 import (
 	"github.com/gin-gonic/gin"
 
-	"github.com/hoophq/hoop/controlplane/backend/internal/apierr"
+	"github.com/hoophq/hoop/controlplane/backend/internal/api/apierr"
 )
 
 const ticket = "EVL-232"
