+++ controlplane/backend/internal/api/server_test.go
@@ -1,4 +1,4 @@
-package httpapi
+package api
 
 import (
 	"io"
