+++ controlplane/backend/internal/api/sidecarauth/sidecarauth.go
@@ -28,7 +28,7 @@ import (
 
 	"github.com/gin-gonic/gin"
 
-	"github.com/hoophq/hoop/controlplane/backend/internal/apierr"
+	"github.com/hoophq/hoop/controlplane/backend/internal/api/apierr"
 )
 
 const ticket = "EVL-234"
